refactor: split AdminAnalyticsSync (694→124) em hook + componentes
Extrai a lógica (3 queries, 8 mutations, runOrdersSync orquestrado, helpers)
para o hook useAnalyticsSync e quebra o JSX em componentes controlados:
SyncEntitiesGrid, ImportCards (clientes/endereços/pedidos), EngineCards
(custo + associação) e EngineConfigCard. Tipos/constantes em types.ts.
Comportamento preservado 1:1; +18 testes.
